application_agent/requirements.txt
telethon
python-dotenv
praw